benchmarks/python/requirements.txt
pyyaml
ruamel.yaml
ruyaml
strictyaml
yamlium
oyaml
ryaml
